Wait for auth response packet
According to the specification here:
https://developer.valvesoftware.com/wiki/Source_RCON_Protocol
there should always be a SERVERDATA_RESPONSE_VALUE coming before
SERVERDATA_AUTH_RESPONSE which is the packet we are waiting for.
This is also the behavior I see when running it with a CS:GO server.
With the current implementation I always get RequestIdMismatch for
any command after auth as the socket requests are out of sync.
This fixes the issue for me.